Move configuration functions to clkConfig.cpp
